Fixes #16

+++ backend/src/main/kotlin/vegait/rs/osipodgorica/service/ImageUploadService.kt
@@ -35,13 +35,12 @@ class ImageUploadService(
         return destinationFile.toString()
     }
 
-    fun deleteFile(path: String): Boolean =
-        File(path).takeIf { file ->
-            file.exists()
-        }?.let { file ->
+    fun deleteFile(path: String) {
+        val file = File(path)
+        if (file.exists()) {
             file.delete()
-            true
-        } ?: false
+        }
+    }
 
     fun deleteFolder(folderName: String) {
         val folder = File(this.rootLocation.resolve(folderName).toString())
